Consensus: MOVEONLY: Move CValidationState from main consensus/validation
Applies: https://github.com/bitcoin/bitcoin/commit/da29ecbcc6c003a603ee2312309d593708ec6fb2
