+++ src/Makefile.am
@@ -99,6 +99,7 @@ BITCOIN_CORE_H = \
   compat/endian.h \
   compat/sanity.h \
   compressor.h \
+  consensus/validation.h \
   core_memusage.h \
   dstencode.h \
   primitives/block.h \
