+++ src/obfuscation.cpp
@@ -5,6 +5,7 @@
 
 #include "obfuscation.h"
 #include "coincontrol.h"
+#include "consensus/validation.h"
 #include "init.h"
 #include "main.h"
 #include "masternodeman.h"
