+++ src/test/transaction_tests.cpp
@@ -8,6 +8,7 @@
 #include "data/tx_valid.json.h"
 
 #include "clientversion.h"
+#include "consensus/validation.h"
 #include "key.h"
 #include "keystore.h"
 #include "main.h"
