+++ src/txmempool.cpp
@@ -8,6 +8,7 @@
 #include "txmempool.h"
 
 #include "clientversion.h"
+#include "consensus/validation.h"
 #include "main.h"
 #include "streams.h"
 #include "util.h"
